+++ idl/base.thrift
@@ -22,10 +22,3 @@ struct rpc_address
 struct gpid
 {
 }
-
-struct request_meta {
-    1:i32 app_id;
-    2:i32 partition_index;
-    3:i32 client_timeout;
-    4:i64 partition_hash;
-}
