+++ idl/replication.thrift
@@ -31,3 +31,10 @@ struct query_cfg_response
     4:bool is_stateful;
     5:list<partition_configuration> partitions;
 }
+
+struct request_meta {
+    1:i32 app_id;
+    2:i32 partition_index;
+    3:i32 client_timeout;
+    4:i64 partition_hash;
+}
